% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_plot_list.R
\name{add_to_plot_list}
\alias{add_to_plot_list}
\title{Add to Plot List}
\usage{
add_to_plot_list(plot_list, plot_or_list)
}
\arguments{
\item{plot_list}{A list to add plots to.}

\item{plot_or_list}{A plot or list of plots to add.}
}
\value{
An updated plot list.
}
\description{
Adds one or more plots to an existing plot list for later export to PowerPoint.
}
