% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_point_plot.R
\name{bar_point_plot}
\alias{bar_point_plot}
\title{Bar plots of grouped data with summary statistics}
\usage{
bar_point_plot(
  d,
  x = NULL,
  y,
  color.group = NULL,
  x.axis = NULL,
  y.axis = NULL,
  legend.title = NULL,
  x.order = NULL,
  group.order = NULL,
  x.first = NULL,
  group.first = NULL,
  points = TRUE,
  barwidth = 0.7,
  jitterwidth = 1,
  pointsize = 1,
  whisker.width = 1,
  mean.type = "arithmetic",
  error.lower = NULL,
  error.upper = NULL
)
}
\arguments{
\item{d}{\code{data.frame} with data to be plotted.}

\item{x}{Character. The name of the column to be used for the x axis (categorical data).
If \code{NULL}, \code{color.group} needs to be supplied. Defaults to \code{NULL}.}

\item{y}{Character. The name of the column to be used for the y axis (numeric data).}

\item{color.group}{Character. The name of the column to be used for color grouping (categorical data).
If \code{NULL},\code{x} needs to be supplied. Defaults to \code{NULL}.}

\item{x.axis}{Character. Sets the title of the x axis. If \code{NULL}, x axis title is given the value of \code{x}.
Defaults to \code{NULL}.}

\item{y.axis}{Character. Sets the title of the y axis. If \code{NULL}, y axis title is given the value of \code{y}.
Defaults to \code{NULL}.}

\item{legend.title}{Character. Sets the title of the legend. If \code{NULL}, legend title is given the value of \code{color.group}.
Defaults to \code{NULL}.}

\item{x.order}{Character vector of length equal to the number of x categories.
Sets the order of categories in the x axis. If \code{NULL}, x categories are ordered alphabetically. Defaults to \code{NULL}.}

\item{group.order}{Character vector of length equal to the number of color grouping categories.
Sets the order of color groups. If \code{NULL}, color groups are ordered alphabetically. Defaults to \code{NULL}.}

\item{x.first}{Character. Places a specific x axis category first. Ignored if \code{x.order} is supplied.
Defaults to \code{NULL}.}

\item{group.first}{Character. Places a specific color group category first. Ignored if \code{group.order} is supplied.
Defaults to \code{NULL}.}

\item{points}{Logical. Sets whether or not to plot individual data points. Defaults to \code{TRUE}.}

\item{barwidth}{Numeric. Sets the width of the bars. Defaults to 0.7.}

\item{jitterwidth}{Numeric. The horizontal dispersion of individual data points. Defaults to 1.}

\item{pointsize}{Numeric. Sets the size of individual data points. Defaults to 1.}

\item{whisker.width}{Numeric. Sets the width of the error bar whiskers. Defaults to 1.}

\item{mean.type}{Character, either "arithmetic" or "geometric". Sets which type of mean and standard deviation to plot. Defaults to "arithmetic".}

\item{error.lower}{Character, name of a column in the \code{d} \code{data.frame}.
Plot custom lower error bar, calculated by the user and included in the \code{d} data frame as a column. Defaults to \code{NULL}, which plots \code{mean - SD}.}

\item{error.upper}{Character, name of a column in the \code{d} \code{data.frame}.
Plot custom upper error bar, calculated by the user and included in the \code{d} data frame as a column. Defaults to \code{NULL}, which plots \code{mean + SD}.}
}
\value{
A plot based on \code{ggplot2}.
}
\description{
Creates a bar plot of grouped data means plus/minus standard deviations (or custom, manually calculated error measure).
}
\details{
The calculations for mean and standard deviation are performed automatically. To show other error measure,
calculate it manually and include it as two fields in the supplied \code{data.frame}
(field names are supplied as \code{error.lower} and \code{error.upper} arguments).
To change from arithmetic mean plus/minus standard deviation to the corresponding geometric parameters, set \code{mean.type}
argument from "arithmetic" to "geometric".

The order of groups shown in the x axis and color groups defaults to alphabetical.
To change it, supply \code{x.order} or \code{group.order} arguments.
These should be character vectors with the desired order of each factor.
Do not include categories that don't exist in the supplied \code{data.frame}.
To bring just one category in first place, supply \code{x.first} or \code{group.first} arguments.
If \code{x.order} is supplied, \code{x.first} will be ignored. The same holds for \code{group.order} and \code{group.first}.

Adjust other supplied arguments to customize the plot aesthetically.
}
