% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_mean_plot.R
\name{box_mean_plot}
\alias{box_mean_plot}
\title{Box plots of grouped data with optional mean value}
\usage{
box_mean_plot(
  d,
  x = NULL,
  y,
  color.group = NULL,
  x.axis = NULL,
  y.axis = NULL,
  legend.title = NULL,
  x.order = NULL,
  group.order = NULL,
  x.first = NULL,
  group.first = NULL,
  means = FALSE,
  boxwidth = 0.7,
  whisker.width = 1,
  mean.size = 1,
  points = TRUE,
  jitterwidth = 1,
  pointsize = 1,
  mean.type = "arithmetic"
)
}
\arguments{
\item{d}{\code{data.frame} with data to be plotted.}

\item{x}{Character. The name of the column to be used for the x axis (categorical data).
Defaults to \code{NULL}, but then \code{color.group} needs to be supplied.}

\item{y}{Character. The name of the column to be used for the y axis (numeric data).}

\item{color.group}{Character. The name of the column to be used for color grouping (categorical data).
Defaults to \code{NULL}, but then \code{x} needs to be supplied.}

\item{x.axis}{Character. The title of the x axis. If \code{NULL}, x axis title is given the value of \code{x}.
Defaults to \code{NULL}.}

\item{y.axis}{Character. The title of the y axis. If \code{NULL}, y axis title is given the value of \code{y}.
Defaults to \code{NULL}.}

\item{legend.title}{Character. The title of the legend. If \code{NULL}, legend title is given the value of \code{color.group}.
Defaults to \code{NULL}.}

\item{x.order}{Character vector of length equal to the number of x categories.
Sets the order of categories in the x axis. Defaults to alphabetical order.}

\item{group.order}{Character vector of length equal to the number of color grouping categories.
Sets the order of color groups. Defaults to alphabetical order.}

\item{x.first}{Character. Places a specific x axis category first. Ignored if \code{x.order} is supplied.
Defaults to \code{NULL}.}

\item{group.first}{Character. Places a specific color group category first.
Ignored if \code{group.order} is supplied. Defaults to \code{NULL}.}

\item{means}{Logical. Sets whether or not to plot group means. Defaults to \code{FALSE}.}

\item{boxwidth}{Numeric. Sets the width of the bars. Defaults to 0.7.}

\item{whisker.width}{Numeric. Sets the width of the error bar whiskers. Defaults to 1.}

\item{mean.size}{Numeric. Sets the size of group mean points. Defaults to 1.}

\item{points}{Logical. Sets whether or not to plot individual data points. Defaults to \code{TRUE}.
Note that, even if set to \code{FALSE}, outliers will still be shown as individual data points, as is the norm with box plots.}

\item{jitterwidth}{Numeric. The horizontal dispersion of individual data points. Defaults to 1.}

\item{pointsize}{Numeric. Sets the size of the individual data points. Defaults to 1.}

\item{mean.type}{Character. Sets from arithmetic to geometric mean and SD. Defaults to "arithmetic".}
}
\value{
A plot based on \code{ggplot2}.
}
\description{
Create standard box plots of grouped data (line: median, box: 25 to 75 percentiles, whiskers: 1.5*IQR (interquartile range)).
}
\details{
Can include the mean value for each group. To include it, set \code{means} argument to \code{TRUE}. To change between arithmetic or geometric
mean, set the \code{mean.type} argument to "arithmetic" or "geometric".

The order of groups shown in the x axis and color groups defaults to alphabetical.
To change it, supply \code{x.order} or \code{group.order} arguments.
These should be character vectors with the desired order of each factor.
Do not include categories that don't exist in the supplied \code{data.frame}.
To bring just one category first, supply \code{x.first} or \code{group.first} arguments.
If \code{x.order} is supplied, \code{x.first} will be ignored. The same holds for \code{group.order} and \code{group.first}.

Adjust other supplied arguments to customize the plot aesthetically.
}
