% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\docType{package}
\name{datavis-package}
\alias{datavis}
\alias{datavis-package}
\title{datavis: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
