% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{idata}
\alias{idata}
\title{Visualize the master files interactively.  Creates a shiny app to
choose the master files to look at with datatables.}
\usage{
idata(update = FALSE)
}
\arguments{
\item{update}{Check for updates to master files (default FALSE).}
}
\description{
Visualize the master files interactively.  Creates a shiny app to
choose the master files to look at with datatables.
}

