% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_save.R
\name{plot_save}
\alias{plot_save}
\title{Plot export in multiple formats}
\usage{
plot_save(plot, filepath, height, width, unit = "in", dpi = 600)
}
\arguments{
\item{plot}{Plot to be exported. Needs to be based on \code{ggplot2} (\code{datavis}-created plots comply with this).}

\item{filepath}{Character. The filepath for the exported plot, relative to the working directory.
Do not add file extension, appropriate extension will be added to each exported file.}

\item{height}{Numeric. The height of the exported plot in the desired length unit.}

\item{width}{Numeric. The width of the exported plot in the desired length unit.}

\item{unit}{Character. The unit to be used as \code{height} and \code{width}.
"in" for inches, "cm" for centimeters and "mm" for millimeters. Defaults to "in".}

\item{dpi}{Integer. The resolution of the exported PNG image in dots per inch. Defaults to 600.}
}
\value{
Exports plot without returning a value.
}
\description{
Exports plots based on \code{ggplot2} in vector (PDF, SVG), raster (PNG) and R data format (Rds).
}
\details{
PNG and PDF files are exported using the \code{ggplot2}-based \code{ggsave} function.
Exported PNG images are anti-aliased using the Cairo graphics library.
Exported SVG graphics are drawn using the R-native SVG graphics device. This is because SVG graphics exported by \code{ggsave}
are not rendered properly by the graphic design software Affinity Designer.
The downside of exporting SVG with the native graphics device is that the text is drawn as paths instead of text,
so it cannot be easily modified as text in graphic design software (e.g. set as bold, italics, etc).
}
