% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stats.R
\name{plot_stats}
\alias{plot_stats}
\title{Add labels on top of bars or boxes}
\usage{
plot_stats(plot, d, labels, position = "identity", size = 1, y.adj = 0)
}
\arguments{
\item{plot}{Plot onto which to add labels. Requires a discrete x scale, optionally dodged by "fill" aesthetic.}

\item{d}{\code{data.frame} containing the labels to be plotted.
Must have categorical fields corresponding to the data used for the original plot.}

\item{labels}{Character. The name of the column with the labels to be plotted.}

\item{position}{Character, either "identity" or "dodge". Sets where (horizontally) the labels should be plotted. Defaults to "identity".
If plot is dodged, "dodge" plots labels above each bar or box. "identity" plots them above the middle of each group of bars or boxes.
If plot is not dodged, position is automatically set to "identity".}

\item{size}{Numeric. Sets the size of labels. Defaults to 1.}

\item{y.adj}{Numeric. Adjusts the height of labels above each bar, box or group. Defaults to 0.
Positive sets it higher, negative sets it lower.}
}
\value{
A plot based on \code{ggplot2}.
}
\description{
Adds text labels on top of bars or boxes of a \code{ggplot2}-based plot.
Useful for annotating statistical analysis information.
}
\details{
\code{plot_stats} detects the maximum plotted value of each data group
(be it bar, box, error bar or individual data point), and plots labels above it.
}
