% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_save.R
\name{plot_save}
\alias{plot_save}
\title{Plot export in multiple formats}
\usage{
plot_save(plot, filepath, height, width, unit = "in", dpi = 600)
}
\arguments{
\item{plot}{Plot to be exported. Needs to have been produced using \code{ggplot2} (functions described in this guide comply with this).}

\item{filepath}{Character. The directory and file name for the exported plot.
Do not add file extension, the appropriate extension will be added to each exported file.}

\item{height}{Numeric. The height of the exported plot in the desired length unit.}

\item{width}{Numeric. The width of the exported plot in the desired length unit.}

\item{unit}{Character. The unit to be used as \code{height} and \code{width}.
"in" for inches, "cm" for centimeters and "mm" for millimeters. Defaults to "in".}

\item{dpi}{Integer. The resolution of the exported PNG image in dots per inch. Defaults to 600.}
}
\value{
Exports plot without returning a value.
}
\description{
Exports plots based on \code{ggplot2} in vector (PDF, SVG), raster (PNG) and R data format (Rds).
}
